// src/interfaces/contact-interface.ts
export interface ICreateContact {
  title?: string; //initials of the contact, e.g:Mr.Mrs.
  fullName: string; //Full name of the contact
  phone: string; // The phone number of the contact
  mobile?: string; // The mobile number of the contact
  email?: string; //contact's valid email address
  homeAddress?: string; //Home address of the contact
  company?: string; //contact's company name
  country?: string; //contact's country
  state?: string; //contact's state
  street?: string; //contact's street name
  zipCode?: string; //contact's country zip code
  website?: string; //contact's website address
}
export interface IContactID {
  id: string; //The ID of the contact
}
export interface IData {
  metadata: IMetadata;
  contact: ICreateContact;
}

export interface IMetadata {
  contactID: string; // The uuid of the  contact
  createdAt: string; // The ISO date of when the contact was created}
  blocked: boolean; //Tells if this contact is Blocked or not
}